Class_02_DACA__OpenRouter_LiteLLM/pyproject.toml
[project]
name = "class-02-daca-openrouter-litellm-with-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents[litellm]>=0.0.13",
    "python-dotenv>=1.1.0",
]
